experiments/requirements.txt
mlflow==2.9.2
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4
prometheus_client==0.19.0
python-dotenv==1.0.0